finished exercise
Successfully completed the exercise.
created method to clean up phone numbers
created method to split an array of date-times into arrays of components
and store as a hash
created method to count each component in an array and return a hash
with appropriately named keys containing the component and the number of
time the component appeared in the array.